Changed int to NSUinteger

+++ UINavigationController-Blocks/UINavigationController-Blocks/UINavigationController-Blocks/MRStack.m
@@ -47,7 +47,7 @@
 	[self.array removeAllObjects];
 }
 
-- (int)count
+- (NSUInteger)count
 {
 	return self.array.count;
 }
